update NatSpec and README

+++ contracts/DelegationManager.sol
@@ -1,13 +1,19 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.28;
 
-import "@openzeppelin/contracts/utils/math/Math.sol";
-import "@openzeppelin/contracts/token/ERC20/utils/SafeERC20.sol";
+import {Math} from "@openzeppelin/contracts/utils/math/Math.sol";
+import {SafeERC20} from "@openzeppelin/contracts/token/ERC20/utils/SafeERC20.sol";
 import {ReentrancyGuard} from "@openzeppelin/contracts/utils/ReentrancyGuard.sol";
-import "@openzeppelin/contracts/token/ERC20/IERC20.sol";
+import {IERC20} from "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 import {IStakingManager} from "./interface/IStakingManager.sol";
 import {Ownable} from "@openzeppelin/contracts/access/Ownable.sol";
 
+/**
+ * @title DelegationManager
+ * @notice Holds delegated staking tokens and distributes time-based rewards to delegators.
+ * @dev Resolver eligibility is checked when users delegate or redelegate, but is not guaranteed
+ * for the lifetime of a delegation. Delegators must monitor resolver eligibility themselves.
+ */
 contract DelegationManager is Ownable, ReentrancyGuard {
     using SafeERC20 for IERC20;
 
@@ -68,6 +74,14 @@ contract DelegationManager is Ownable, ReentrancyGuard {
         _;
     }
 
+    /**
+     * @notice Creates a delegation and reward distribution manager.
+     * @param _rewardsToken ERC-20 token paid as delegation rewards.
+     * @param _stakingToken ERC-20 token deposited by delegators.
+     * @param _stakingManager Contract used to check resolver eligibility.
+     * @param _durationInDays Duration of each reward period, in days.
+     * @param _freezingPeriod Delay in seconds before a delegation can be withdrawn or redelegated.
+     */
     constructor(
         address _rewardsToken,
         address _stakingToken,
@@ -95,14 +109,27 @@ contract DelegationManager is Ownable, ReentrancyGuard {
         freezingPeriod = _freezingPeriod;
     }
 
+    /**
+     * @notice Returns an account's currently delegated token amount.
+     * @param delegator Account whose delegation is queried.
+     * @return amount Current delegated amount.
+     */
     function delegatedAmountOf(address delegator) external view returns (uint256) {
         return delegations[delegator].delegatedAmount;
     }
 
+    /**
+     * @notice Returns the latest timestamp at which rewards may accrue.
+     * @return timestamp The lesser of the current timestamp and reward-period end.
+     */
     function lastTimeRewardApplicable() public view returns (uint256) {
         return Math.min(block.timestamp, durationEnd);
     }
 
+    /**
+     * @notice Returns cumulative rewards per delegated token, scaled by 1e18.
+     * @return accumulatedRewardPerToken Current reward-per-token accumulator.
+     */
     function rewardPerToken() public view returns (uint256) {
         if (totalDelegated == 0) {
             return rewardPerTokenStored;
@@ -112,6 +139,11 @@ contract DelegationManager is Ownable, ReentrancyGuard {
             ((lastTimeRewardApplicable() - lastUpdateTime) * rewardRate * 1e18) / totalDelegated;
     }
 
+    /**
+     * @notice Returns the accrued, unclaimed reward for a delegator.
+     * @param delegator Account whose reward is queried.
+     * @return reward Amount currently accrued for the account.
+     */
     function earned(address delegator) public view returns (uint256) {
         return
             (delegations[delegator].delegatedAmount *
@@ -120,10 +152,21 @@ contract DelegationManager is Ownable, ReentrancyGuard {
             rewards[delegator];
     }
 
+    /**
+     * @notice Returns the scheduled rewards for one complete reward period.
+     * @return reward Total reward emitted during the period at the current rate.
+     */
     function getRewardForDuration() public view returns (uint256) {
         return rewardRate * rewardsDuration;
     }
 
+    /**
+     * @notice Delegates staking tokens to an eligible resolver.
+     * @dev Eligibility is checked only for this call. Delegators must monitor the resolver after
+     * delegating because its self-stake may later fall below the eligibility threshold.
+     * @param resolver Resolver receiving delegation attribution.
+     * @param delegationAmount Amount of staking tokens to delegate.
+     */
     function delegate(
         address resolver,
         uint256 delegationAmount
@@ -153,6 +196,10 @@ contract DelegationManager is Ownable, ReentrancyGuard {
         emit Delegated(delegator, resolver, delegationAmount);
     }
 
+    /**
+     * @notice Withdraws some or all of the caller's delegation after the freezing period.
+     * @param undelegationAmount Amount of staking tokens to withdraw.
+     */
     function undelegate(uint256 undelegationAmount) public nonReentrant updateReward(msg.sender) {
         if (undelegationAmount == 0) revert InvalidUndelegationAmount();
         Delegation storage delegationInfo = delegations[msg.sender];
@@ -180,6 +227,11 @@ contract DelegationManager is Ownable, ReentrancyGuard {
         emit Undelegated(delegator, undelegationAmount);
     }
 
+    /**
+     * @notice Moves the caller's full delegation to another eligible resolver.
+     * @dev A successful redelegation resets the freezing period.
+     * @param newResolver Resolver receiving the delegation attribution.
+     */
     function redelegate(address newResolver) external nonReentrant updateReward(msg.sender) {
         if (!STAKING_MANAGER.isValidResolver(newResolver)) revert NotValidResolver();
 
@@ -200,6 +252,7 @@ contract DelegationManager is Ownable, ReentrancyGuard {
         emit Redelegated(msg.sender, oldResolver, newResolver);
     }
 
+    /** @notice Claims all rewards accrued by the caller. */
     function claimReward() public nonReentrant updateReward(msg.sender) {
         _claimReward(msg.sender);
     }
@@ -213,6 +266,7 @@ contract DelegationManager is Ownable, ReentrancyGuard {
         }
     }
 
+    /** @notice Withdraws the caller's entire delegation and claims all accrued rewards. */
     function exit() external nonReentrant updateReward(msg.sender) {
         Delegation storage delegationInfo = delegations[msg.sender];
         _undelegate(msg.sender, delegationInfo.delegatedAmount, delegationInfo);
@@ -220,10 +274,10 @@ contract DelegationManager is Ownable, ReentrancyGuard {
     }
 
     /**
-     * @notice Notify the contract about a new reward amount.
-     * @dev This function is called by the owner to notify the contract about a new reward amount.
-     * @dev Owner should transfer staking tokens and then call this function to notify the contract about the new reward amount upon contract deployment.
-     * @param reward The amount of reward tokens to be distributed.
+     * @notice Starts or replaces the reward schedule using a new reward amount.
+     * @dev The owner must transfer sufficient reward tokens to this contract before calling this
+     * function; this function does not transfer tokens from the owner.
+     * @param reward Amount of reward tokens to distribute over the next period.
      */
     function notifyRewardAmount(uint256 reward) external onlyOwner updateReward(address(0)) {
         if (reward == 0) revert InvalidRewardAmount();
@@ -239,7 +293,7 @@ contract DelegationManager is Ownable, ReentrancyGuard {
         // This keeps the reward rate in the right range, preventing overflows due to
         // very high values of rewardRate in the earned and rewardsPerToken functions;
         // Reward + leftover must be less than 2^256 / 10^18 to avoid overflow.
-        uint balance = REWARD_TOKEN.balanceOf(address(this));
+        uint256 balance = REWARD_TOKEN.balanceOf(address(this));
         if (rewardRate > balance / rewardsDuration) revert RewardAmountTooHigh();
 
         lastUpdateTime = block.timestamp;
@@ -247,15 +301,29 @@ contract DelegationManager is Ownable, ReentrancyGuard {
         emit RewardAdded(reward);
     }
 
+    /**
+     * @notice Updates the delay before delegation withdrawal or redelegation.
+     * @dev Trusted owner governance may set this value to zero.
+     * @param _freezingPeriod New freezing period in seconds.
+     */
     function setFreezingPeriod(uint256 _freezingPeriod) external onlyOwner {
         freezingPeriod = _freezingPeriod;
         emit FreezingPeriodUpdated(freezingPeriod);
     }
 
+    /**
+     * @notice Returns whether an account has a non-zero active delegation.
+     * @param delegator Account whose delegation is queried.
+     * @return valid True when the account has an active delegation.
+     */
     function isValidDelegation(address delegator) public view returns (bool) {
         return delegations[delegator].delegatedAmount > 0;
     }
 
+    /**
+     * @notice Returns total staking tokens held as delegations.
+     * @return amount Aggregate delegated amount.
+     */
     function getTotalDelegated() external view returns (uint256) {
         return totalDelegated;
     }
